fix(route): split openai common logic to common.js
