+++ lib/v2/openai/blog.js
@@ -1,8 +1,6 @@
 const got = require('@/utils/got');
-const cheerio = require('cheerio');
-const { art } = require('@/utils/render');
 const { toTitleCase } = require('@/utils/common-utils');
-const path = require('path');
+const { getApiUrl, parseArticle } = require('./common');
 
 module.exports = async (ctx) => {
     const tag = ctx.params.tag || '';
@@ -11,18 +9,7 @@ module.exports = async (ctx) => {
     const blogRootUrl = 'https://openai.com/blog';
     const blogOriginUrl = `${rootUrl}/blog${tag === '' ? '' : `?topics=${tag}`}`;
 
-    // Find API base URL
-    const initResponse = await got({
-        method: 'get',
-        url: blogRootUrl,
-    });
-
-    let apiBaseUrl = initResponse.data
-        .toString()
-        .match(/(?<=TWILL_API_BASE:").+?(?=")/)[0]
-        .replaceAll('\\u002F', '/');
-    apiBaseUrl = apiBaseUrl + '/api/v1/blog-details';
-    const apiUrl = new URL(apiBaseUrl);
+    const apiUrl = new URL((await getApiUrl()) + '/api/v1/blog-details');
 
     // Construct API query
     apiUrl.searchParams.append('sort', '-publicationDate,-createdAt');
@@ -43,47 +30,7 @@ module.exports = async (ctx) => {
     const items = await Promise.all(
         list.map((item) => {
             const attributes = item.attributes;
-            const textUrl = `${blogRootUrl}/${attributes.slug}`;
-            return ctx.cache.tryGet(attributes.slug, async () => {
-                const detailResponse = await got({
-                    method: 'get',
-                    url: textUrl,
-                });
-                let content = cheerio.load(detailResponse.data);
-
-                const authors = content('[aria-labelledby="metaAuthorsHeading"] > li > a > span > span')
-                    .toArray()
-                    .map((entry) => content(entry).text())
-                    .join(', ');
-
-                // Leave out comments
-                const comments = content('*')
-                    .contents()
-                    .filter(function () {
-                        return this.nodeType === 8;
-                    });
-                comments.remove();
-
-                content = content('#content');
-
-                const imageSrc = attributes.seo.ogImageSrc;
-                const imageAlt = attributes.seo.ogImageAlt;
-
-                const article = art(path.join(__dirname, 'templates/article.art'), {
-                    content,
-                    imageSrc,
-                    imageAlt,
-                });
-
-                return {
-                    title: attributes.title,
-                    author: authors,
-                    description: article,
-                    pubDate: attributes.createdAt,
-                    category: attributes.tags.map((tag) => tag.title),
-                    link: textUrl,
-                };
-            });
+            return parseArticle(ctx, blogRootUrl, attributes);
         })
     );
