+++ lib/v2/openai/maintainer.js
@@ -1,3 +1,4 @@
 module.exports = {
     '/blog/:tag?': ['StevenRCE0', 'nczitzk'],
+    '/research': ['yuguorui'],
 };
